+++ ironfish/src/mining/soloMiner.ts
@@ -7,7 +7,7 @@ import { Assert } from '../assert'
 import { Logger } from '../logger'
 import { Meter } from '../metrics/meter'
 import { Target } from '../primitives/target'
-import { IronfishIpcClient } from '../rpc/clients/ipcClient'
+import { IronfishRpcClient } from '../rpc/clients/rpcClient'
 import { SerializedBlockTemplate } from '../serde/BlockTemplateSerde'
 import { BigIntUtils } from '../utils/bigint'
 import { ErrorUtils } from '../utils/error'
@@ -22,7 +22,7 @@ export class MiningSoloMiner {
   readonly hashRate: Meter
   readonly threadPool: ThreadPoolHandler
   readonly logger: Logger
-  readonly rpc: IronfishIpcClient
+  readonly rpc: IronfishRpcClient
 
   private started: boolean
   private stopPromise: Promise<void> | null
@@ -48,7 +48,7 @@ export class MiningSoloMiner {
     batchSize: number
     logger: Logger
     graffiti: Buffer
-    rpc: IronfishIpcClient
+    rpc: IronfishRpcClient
   }) {
     this.rpc = options.rpc
     this.logger = options.logger
@@ -240,7 +240,7 @@ export class MiningSoloMiner {
     if (!connected) {
       if (!this.connectWarned) {
         this.logger.warn(
-          `Failed to connect to node on ${String(this.rpc.connection.mode)}, retrying...`,
+          `Failed to connect to node on ${String(this.rpc.connectionMode)}, retrying...`,
         )
         this.connectWarned = true
       }
