+++ ironfish/src/rpc/clients/index.ts
@@ -6,3 +6,4 @@ export * from './client'
 export * from './ipcClient'
 export * from './memoryClient'
 export * from './rpcClient'
+export * from './tcpClient'
